requirements.txt
altair
appnope==0.1.4
asttokens==2.4.1
attrs
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Bottleneck
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
gitdb==4.0.11
GitPython==3.1.43
htbuilder==0.6.2
idna==3.7
importlib-metadata
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema
jsonschema-specifications
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numexpr
numpy
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.22.0
plotly-express==0.4.1
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pydeck==0.9.0
Pygments==2.18.0
python-dateutil
pytz
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py
scipy==1.13.0
six
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.34.0
tenacity==8.3.0
text-highlighter==0.0.11
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp
